+++ src/tls/mod.rs
@@ -1,7 +1,5 @@
 #[cfg(feature = "sources-tls")]
-use futures01::{try_ready, Async, Future, Poll, Stream};
-#[cfg(feature = "sources-tls")]
-use openssl::ssl::{HandshakeError, SslAcceptor};
+use openssl::ssl::HandshakeError;
 use openssl::{
     error::ErrorStack,
     pkcs12::{ParsedPkcs12, Pkcs12},
@@ -17,19 +15,17 @@ use snafu::{ResultExt, Snafu};
 use std::fmt::{self, Debug, Formatter};
 use std::fs::File;
 use std::io::Read;
-#[cfg(feature = "sources-tls")]
-use std::net::SocketAddr;
 use std::path::{Path, PathBuf};
 #[cfg(feature = "sources-tls")]
-use tokio::net::{tcp::Incoming, TcpListener, TcpStream};
-#[cfg(feature = "sources-tls")]
-use tokio_openssl::{AcceptAsync, SslAcceptorExt, SslStream};
+use tokio::net::TcpStream;
 
+#[cfg(any(feature = "sources-tls", feature = "sources-http"))]
+mod incoming;
 mod maybe_tls;
-pub use maybe_tls::MaybeTls;
 
-#[cfg(feature = "sources-tls")]
-pub type MaybeTlsStream<S> = MaybeTls<S, SslStream<S>>;
+#[cfg(any(feature = "sources-tls", feature = "sources-http"))]
+pub(crate) use incoming::MaybeTlsSettings;
+pub(crate) use maybe_tls::MaybeTls;
 
 #[derive(Debug, Snafu)]
 pub enum TlsError {
@@ -245,19 +241,6 @@ impl TlsSettings {
         })
     }
 
-    #[cfg(feature = "sources-tls")]
-    pub(crate) fn acceptor(&self) -> crate::Result<SslAcceptor> {
-        match self.identity {
-            None => Err(TlsError::MissingRequiredIdentity.into()),
-            Some(_) => {
-                let mut acceptor =
-                    SslAcceptor::mozilla_intermediate(SslMethod::tls()).context(CreateAcceptor)?;
-                self.apply_context(&mut acceptor)?;
-                Ok(acceptor.build())
-            }
-        }
-    }
-
     fn apply_context(&self, context: &mut SslContextBuilder) -> crate::Result<()> {
         context.set_verify(if self.verify_certificate {
             SslVerifyMode::PEER | SslVerifyMode::FAIL_IF_NO_PEER_CERT
@@ -324,34 +307,6 @@ impl Debug for TlsSettings {
     }
 }
 
-#[cfg(feature = "sources-tls")]
-pub(crate) type MaybeTlsSettings = MaybeTls<(), TlsSettings>;
-
-#[cfg(feature = "sources-tls")]
-impl MaybeTlsSettings {
-    pub(crate) fn from_config(
-        config: &Option<TlsConfig>,
-        require_ident: bool,
-    ) -> crate::Result<Self> {
-        Ok(match TlsSettings::from_config(config, require_ident)? {
-            None => Self::Raw(()),
-            Some(tls) => Self::Tls(tls),
-        })
-    }
-
-    pub(crate) fn bind(&self, addr: &SocketAddr) -> crate::Result<MaybeTlsIncoming<Incoming>> {
-        let listener = TcpListener::bind(addr)?;
-        let incoming = listener.incoming();
-
-        let acceptor = match self {
-            Self::Tls(tls) => Some(tls.acceptor()?.into()),
-            Self::Raw(()) => None,
-        };
-
-        MaybeTlsIncoming::new(incoming, acceptor)
-    }
-}
-
 /// Load a private key from a named file
 fn load_key(filename: &Path, pass_phrase: &Option<String>) -> crate::Result<PKey<Private>> {
     let data = open_read(filename, "key")?;
@@ -384,80 +339,6 @@ fn open_read(filename: &Path, note: &'static str) -> crate::Result<Vec<u8>> {
         .with_context(|| FileReadFailed { note, filename })?;
 
     Ok(text)
-}
-
-#[cfg(feature = "sources-tls")]
-pub(crate) struct MaybeTlsIncoming<I: Stream> {
-    incoming: I,
-    acceptor: Option<SslAcceptor>,
-    state: MaybeTlsIncomingState<I::Item>,
-}
-
-#[cfg(feature = "sources-tls")]
-enum MaybeTlsIncomingState<S> {
-    Inner,
-    Accepting(AcceptAsync<S>),
-}
-
-#[cfg(feature = "sources-tls")]
-impl<I: Stream> MaybeTlsIncoming<I> {
-    pub(crate) fn new(incoming: I, acceptor: Option<SslAcceptor>) -> crate::Result<Self> {
-        Ok(Self {
-            incoming,
-            acceptor,
-            state: MaybeTlsIncomingState::Inner,
-        })
-    }
-}
-
-#[cfg(feature = "sources-tls")]
-impl Stream for MaybeTlsIncoming<Incoming> {
-    type Item = MaybeTlsStream<TcpStream>;
-    type Error = TlsError;
-
-    fn poll(&mut self) -> Poll<Option<Self::Item>, Self::Error> {
-        loop {
-            match &mut self.state {
-                MaybeTlsIncomingState::Inner => {
-                    let stream = if let Some(stream) = try_ready!(self
-                        .incoming
-                        .poll()
-                        .map_err(Into::into)
-                        .context(IncomingListener))
-                    {
-                        stream
-                    } else {
-                        return Ok(Async::Ready(None));
-                    };
-
-                    if let Some(acceptor) = &mut self.acceptor {
-                        let fut = acceptor.accept_async(stream);
-
-                        self.state = MaybeTlsIncomingState::Accepting(fut);
-                        continue;
-                    } else {
-                        return Ok(Async::Ready(Some(MaybeTlsStream::Raw(stream))));
-                    }
-                }
-
-                MaybeTlsIncomingState::Accepting(fut) => {
-                    let stream = try_ready!(fut.poll().context(Handshake));
-                    self.state = MaybeTlsIncomingState::Inner;
-
-                    return Ok(Async::Ready(Some(MaybeTlsStream::Tls(stream))));
-                }
-            }
-        }
-    }
-}
-
-#[cfg(feature = "sources-tls")]
-impl<I: Stream + Debug> Debug for MaybeTlsIncoming<I> {
-    fn fmt(&self, f: &mut Formatter) -> fmt::Result {
-        f.debug_struct("MaybeTlsIncoming")
-            .field("incoming", &self.incoming)
-            .finish()
-    }
 }
 
 #[cfg(test)]
